moved code-generator files to a dedicated dir
