+++ Makefile
@@ -11,7 +11,7 @@ test: src/lexical-analyser.test.c src/states-machine.test.c src/lexical-analyser
 	@gcc -o dist/states-machine.test src/states-machine.test.c src/states-machine.c -I.
 	@gcc -o dist/lexical-analyser.test src/lexical-analyser.test.c src/lexical-analyser.c src/token.c src/states-machine.c -I.
 	@gcc -o dist/syntatic-analyser.test src/syntatic-analyser.test.c src/syntatic-analyser.c src/code-generator/code-generator-transition.c src/create_ape.c src/ape.c src/lexical-analyser.c src/token.c src/states-machine.c -I.
-	@gcc -o dist/code-generator.test src/code-generator.test.c src/code-generator.c src/code-generator/code-generator-transition.c src/syntatic-analyser.c src/create_ape.c src/ape.c src/lexical-analyser.c src/token.c src/states-machine.c -I.
+	@gcc -o dist/code-generator.test src/code-generator/code-generator.test.c src/code-generator/code-generator.c src/code-generator/code-generator-transition.c src/syntatic-analyser.c src/create_ape.c src/ape.c src/lexical-analyser.c src/token.c src/states-machine.c -I.
 	@./dist/states-machine.test
 	@./dist/lexical-analyser.test src/fixtures/hello.el
 	@./dist/syntatic-analyser.test
