+++ src/code-generator/code-generator.h
@@ -1,4 +1,4 @@
-#include "./token.h"
+#include "../token.h"
 
 void consumeTransition(Token *token, char *submachine, char *state);
 void initCodeGenerator();
